Started implementation of rtaudio for playback, and changed all usages of double primitives to floats in oscillator, synthesizer, and envelopes

+++ Synthesizer/Envelopes/Decorators/EnvelopeDecorator.cpp
@@ -1,6 +1,6 @@
 #include "EnvelopeDecorator.h"
 
-EnvelopeDecorator::EnvelopeDecorator(std::unique_ptr<Envelope> env, double sample_rate) {
+EnvelopeDecorator::EnvelopeDecorator(std::unique_ptr<Envelope> env, float sample_rate) {
     this->base_envelope = std::move(env);
     this->sample_rate = sample_rate;
 }
@@ -15,7 +15,7 @@ void EnvelopeDecorator::off() {
     base_envelope->off();
 }
 
-double EnvelopeDecorator::get_multiplier() {
+float EnvelopeDecorator::get_multiplier() {
     return base_envelope->get_multiplier();
 }
