+++ Synthesizer/Envelopes/Decorators/TremoloEnvelope.h
@@ -3,22 +3,21 @@
 #include <memory>
 #include "EnvelopeDecorator.h"
 
-constexpr double TWO_PI = 2 * 3.14159265358979323846;
+constexpr float TWO_PI = 2 * 3.14159265358979323846;
 
 class TremoloEnvelope : public EnvelopeDecorator {
 private:
-    double current_phase;
-    double phase_increment;
+    float current_phase;
+    float phase_increment;
 
-    double speed_hz;
-    double depth;
+    float speed_hz;
+    float depth;
 
 public:
-    TremoloEnvelope(std::unique_ptr<Envelope> env, double sample_rate = 44100.0,
-        double speed_hz = 5.0, double depth = 0.5);
+    TremoloEnvelope(std::unique_ptr<Envelope> env, float sample_rate = 44100.0, float speed_hz = 5.0, float depth = 0.5);
     ~TremoloEnvelope();
 
-    double get_multiplier() override;
+    float get_multiplier() override;
 };
