+++ Synthesizer/Oscillators/Decorators/VibratoOscillator.cpp
@@ -2,7 +2,7 @@
 
 #include <cmath>
 
-VibratoOscillator::VibratoOscillator(std::unique_ptr<Oscillator> osc, double sample_rate, double base_hz, double speed_hz, double depth)
+VibratoOscillator::VibratoOscillator(std::unique_ptr<Oscillator> osc, float sample_rate, float base_hz, float speed_hz, float depth)
         : OscillatorDecorator(std::move(osc), sample_rate) {
     this->base_hz = base_hz;
     this->speed_hz = speed_hz;
@@ -13,16 +13,16 @@ VibratoOscillator::VibratoOscillator(std::unique_ptr<Oscillator> osc, double sam
 
 VibratoOscillator::~VibratoOscillator() = default;
 
-void VibratoOscillator::set_frequency(double hz, double sample_rate) {
+void VibratoOscillator::set_frequency(float hz, float sample_rate) {
     base_hz = hz;
     OscillatorDecorator::set_frequency(hz, sample_rate);
 }
 
-double VibratoOscillator::get_sample() {
-    double lfo_wobble = std::sin(current_phase);
+float VibratoOscillator::get_sample() {
+    float lfo_wobble = std::sin(current_phase);
     current_phase = fmod(current_phase + phase_increment, TWO_PI);
 
-    double pitch_offset = lfo_wobble * depth;
+    float pitch_offset = lfo_wobble * depth;
     base_oscillator->set_frequency(base_hz + pitch_offset, sample_rate);
     return base_oscillator->get_sample();
